Finish the core of `/feed` route
Now, we are able to get all the posts with images and captions.
Posts will show up from newest to latest
Username and profile pic of the person who posted are also returned
Results are returned as JSON

Future plan:
Work on intergrate user ID from the auth part to this part
Look into how to use this endpoint to the front end. Contact Danny for it.